Iteration 2
Redesign Changes, Ternary Relationship Design Where the 3 Entities: 
Patient, Referrer and Surgeon are all linked to the Referral Bridging Table.
FSA has been moved to the Referral Table as an Attribute instead of being an Entity. 
Addressing Foreign key Attributes to the Referral Table so reverse engineering the diagram on SSMS will be correct.